+++ @types/globals.d.ts
@@ -103,3 +103,5 @@ declare type Setter = (value: any) => void;
 declare const Buffer: any;
 
 declare type EnumAlias<EnumT> = EnumT[keyof EnumT];
+
+declare module 'internal:native' {}
