+++ cocos/core/assets/rendering-sub-mesh.jsb.ts
@@ -56,6 +56,7 @@ export const RenderingSubMesh: typeof JsbRenderingSubMesh = jsb.RenderingSubMesh
 
 const renderingSubMeshProto = RenderingSubMesh.prototype;
 
+// @ts-expect-error Property '_ctor' does not exist on type 'RenderingSubMesh'.
 renderingSubMeshProto._ctor = function (vertexBuffers: Buffer[], attributes: Attribute[], primitiveMode: PrimitiveMode,
     indexBuffer: Buffer | null = null, indirectBuffer: Buffer | null = null) {
     jsb.Asset.prototype._ctor.apply(this, arguments);
