+++ cocos/core/assets/texture-2d.jsb.ts
@@ -51,6 +51,7 @@ export interface ITexture2DSerializeData {
 }
 
 texture2DProto._ctor = function () {
+    // @ts-expect-error Property '_ctor' does not exist on type 'SimpleTexture'.
     SimpleTexture.prototype._ctor.apply(this, arguments);
     this._mipmaps = [];
 };
