+++ cocos/core/math/quat.ts
@@ -689,7 +689,7 @@ export class Quat extends ValueType {
 
     constructor (x?: number | IQuatLike, y?: number, z?: number, w?: number) {
         super();
-        if (x && typeof x === 'object') {
+        if (typeof x === 'object') {
             this.x = x.x;
             this.y = x.y;
             this.z = x.z;
@@ -726,7 +726,7 @@ export class Quat extends ValueType {
     public set (x?: number, y?: number, z?: number, w?: number): Quat;
 
     public set (x?: number | Quat, y?: number, z?: number, w?: number) {
-        if (x && typeof x === 'object') {
+        if (typeof x === 'object') {
             this.x = x.x;
             this.y = x.y;
             this.z = x.z;
