+++ cocos/core/math/vec2.ts
@@ -447,7 +447,7 @@ export class Vec2 extends ValueType {
 
     constructor (x?: number | Vec2, y?: number) {
         super();
-        if (x && typeof x === 'object') {
+        if (typeof x === 'object') {
             this.x = x.x;
             this.y = x.y;
         } else {
@@ -482,7 +482,7 @@ export class Vec2 extends ValueType {
     public set (x?: number, y?: number): Vec2;
 
     public set (x?: number | Vec2, y?: number) {
-        if (x && typeof x === 'object') {
+        if (typeof x === 'object') {
             this.x = x.x;
             this.y = x.y;
         } else {
