+++ cocos/core/platform/sys.ts
@@ -290,7 +290,7 @@ export const sys = {
     /**
      * @internal
      */
-    init () {
+    init (this: any) {
         try {
             let localStorage: Storage | null = sys.localStorage = window.localStorage;
             localStorage.setItem('storage', '');
@@ -313,7 +313,6 @@ export const sys = {
             this.__isWebIOS14OrIPadOS14Env = (sys.os === OS.IOS || sys.os === OS.OSX) && GameGlobal?.isIOSHighPerformanceMode
             && /(OS 1((4\.[0-9])|(5\.[0-3])))|(Version\/1((4\.[0-9])|(5\.[0-3])))/.test(window.navigator.userAgent);
         } else {
-            // @ts-expect-error HACK: this private property only needed on web & wechat JIT
             this.__isWebIOS14OrIPadOS14Env = (sys.os === OS.IOS || sys.os === OS.OSX) && systemInfo.isBrowser
             && /(OS 14)|(Version\/14)/.test(window.navigator.userAgent);
         }
