+++ cocos/core/renderer/scene/camera.jsb.ts
@@ -126,6 +126,7 @@ Object.defineProperty(Camera, "standardExposureValue", {
     configurable: true,
     enumerable: true,
     get () {
+        // @ts-expect-error Property 'getStandardExposureValue' does not exist on type 'typeof Camera'.
         return Camera.getStandardExposureValue();
     },
 });
@@ -134,6 +135,7 @@ Object.defineProperty(Camera, "standardLightMeterScale", {
     configurable: true,
     enumerable: true,
     get () {
+        // @ts-expect-error Property 'getStandardLightMeterScale' does not exist on type 'typeof Camera'.
         return Camera.getStandardLightMeterScale();
     },
 });
