+++ cocos/core/scene-graph/scene-globals.jsb.ts
@@ -316,21 +316,21 @@ const typeDescriptor = Object.getOwnPropertyDescriptor(ShadowsInfoProto, 'type')
 type(ShadowType)(ShadowsInfoProto, 'type', typeDescriptor);
 editable(ShadowsInfoProto, 'type', typeDescriptor);
 const shadowColorDescriptor = Object.getOwnPropertyDescriptor(ShadowsInfoProto, 'shadowColor');
-visible(function (this) { /* Need to copy source code */ })(ShadowsInfoProto, 'shadowColor', shadowColorDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(ShadowsInfoProto, 'shadowColor', shadowColorDescriptor);
 const planeDirectionDescriptor = Object.getOwnPropertyDescriptor(ShadowsInfoProto, 'planeDirection');
 tooltip('i18n:shadow.planeDirection')(ShadowsInfoProto, 'planeDirection', planeDirectionDescriptor);
-visible(function (this) { /* Need to copy source code */ })(ShadowsInfoProto, 'planeDirection', planeDirectionDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(ShadowsInfoProto, 'planeDirection', planeDirectionDescriptor);
 const planeHeightDescriptor = Object.getOwnPropertyDescriptor(ShadowsInfoProto, 'planeHeight');
 tooltip('i18n:shadow.planeHeight')(ShadowsInfoProto, 'planeHeight', planeHeightDescriptor);
-visible(function (this) { /* Need to copy source code */ })(ShadowsInfoProto, 'planeHeight', planeHeightDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(ShadowsInfoProto, 'planeHeight', planeHeightDescriptor);
 type(CCFloat)(ShadowsInfoProto, 'planeHeight', planeHeightDescriptor);
 editable(ShadowsInfoProto, 'planeHeight', planeHeightDescriptor);
 const maxReceivedDescriptor = Object.getOwnPropertyDescriptor(ShadowsInfoProto, 'maxReceived');
-visible(function (this) { /* Need to copy source code */ })(ShadowsInfoProto, 'maxReceived', maxReceivedDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(ShadowsInfoProto, 'maxReceived', maxReceivedDescriptor);
 tooltip('i18n:shadow.maxReceived')(ShadowsInfoProto, 'maxReceived', maxReceivedDescriptor);
 type(CCInteger)(ShadowsInfoProto, 'maxReceived', maxReceivedDescriptor);
 const shadowMapSizeDescriptor = Object.getOwnPropertyDescriptor(ShadowsInfoProto, 'shadowMapSize');
-visible(function (this) { /* Need to copy source code */ })(ShadowsInfoProto, 'shadowMapSize', shadowMapSizeDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(ShadowsInfoProto, 'shadowMapSize', shadowMapSizeDescriptor);
 tooltip('i18n:shadow.shadowMapSize')(ShadowsInfoProto, 'shadowMapSize', shadowMapSizeDescriptor);
 type(ShadowSize)(ShadowsInfoProto, 'shadowMapSize', shadowMapSizeDescriptor);
 ccclass('cc.ShadowsInfo')(ShadowsInfo);
@@ -368,33 +368,33 @@ slide(FogInfoProto, 'fogDensity', fogDensityDescriptor);
 rangeStep(0.01)(FogInfoProto, 'fogDensity', fogDensityDescriptor);
 range([0, 1])(FogInfoProto, 'fogDensity', fogDensityDescriptor);
 type(CCFloat)(FogInfoProto, 'fogDensity', fogDensityDescriptor);
-visible(function (this) { /* Need to copy source code */ })(FogInfoProto, 'fogDensity', fogDensityDescriptor);
+visible(function (this) { /* Need to copy source code */ return false; })(FogInfoProto, 'fogDensity', fogDensityDescriptor);
 const fogStartDescriptor = Object.getOwnPropertyDescriptor(FogInfoProto, 'fogStart');
 tooltip('i18n:fog.fogStart')(FogInfoProto, 'fogStart', fogStartDescriptor);
 rangeStep(0.01)(FogInfoProto, 'fogStart', fogStartDescriptor);
 type(CCFloat)(FogInfoProto, 'fogStart', fogStartDescriptor);
-visible(function (this) { /* Need to copy source code */ })(FogInfoProto, 'fogStart', fogStartDescriptor);
+visible(function (this) { /* Need to copy source code */ return false; })(FogInfoProto, 'fogStart', fogStartDescriptor);
 const fogEndDescriptor = Object.getOwnPropertyDescriptor(FogInfoProto, 'fogEnd');
 tooltip('i18n:fog.fogEnd')(FogInfoProto, 'fogEnd', fogEndDescriptor);
 rangeStep(0.01)(FogInfoProto, 'fogEnd', fogEndDescriptor);
 type(CCFloat)(FogInfoProto, 'fogEnd', fogEndDescriptor);
-visible(function (this) { /* Need to copy source code */ })(FogInfoProto, 'fogEnd', fogEndDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(FogInfoProto, 'fogEnd', fogEndDescriptor);
 const fogAttenDescriptor = Object.getOwnPropertyDescriptor(FogInfoProto, 'fogAtten');
 tooltip('i18n:fog.fogAtten')(FogInfoProto, 'fogAtten', fogAttenDescriptor);
 rangeStep(0.01)(FogInfoProto, 'fogAtten', fogAttenDescriptor);
 rangeMin(0.01)(FogInfoProto, 'fogAtten', fogAttenDescriptor);
 type(CCFloat)(FogInfoProto, 'fogAtten', fogAttenDescriptor);
-visible(function (this) { /* Need to copy source code */ })(FogInfoProto, 'fogAtten', fogAttenDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(FogInfoProto, 'fogAtten', fogAttenDescriptor);
 const fogTopDescriptor = Object.getOwnPropertyDescriptor(FogInfoProto, 'fogTop');
 tooltip('i18n:fog.fogTop')(FogInfoProto, 'fogTop', fogTopDescriptor);
 rangeStep(0.01)(FogInfoProto, 'fogTop', fogTopDescriptor);
 type(CCFloat)(FogInfoProto, 'fogTop', fogTopDescriptor);
-visible(function (this) { /* Need to copy source code */ })(FogInfoProto, 'fogTop', fogTopDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(FogInfoProto, 'fogTop', fogTopDescriptor);
 const fogRangeDescriptor = Object.getOwnPropertyDescriptor(FogInfoProto, 'fogRange');
 tooltip('i18n:fog.fogRange')(FogInfoProto, 'fogRange', fogRangeDescriptor);
 rangeStep(0.01)(FogInfoProto, 'fogRange', fogRangeDescriptor);
 type(CCFloat)(FogInfoProto, 'fogRange', fogRangeDescriptor);
-visible(function (this) { /* Need to copy source code */ })(FogInfoProto, 'fogRange', fogRangeDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(FogInfoProto, 'fogRange', fogRangeDescriptor);
 ccclass('cc.FogInfo')(FogInfo);
 
 const SkyboxInfoProto = SkyboxInfo.prototype;
@@ -435,7 +435,7 @@ displayOrder(100)(SkyboxInfoProto, 'diffuseMap', diffuseMapDescriptor);
 type(TextureCube)(SkyboxInfoProto, 'diffuseMap', diffuseMapDescriptor);
 readOnly(SkyboxInfoProto, 'diffuseMap', diffuseMapDescriptor);
 editable(SkyboxInfoProto, 'diffuseMap', diffuseMapDescriptor);
-visible(function (this) { /* Need to copy source code */ })(SkyboxInfoProto, 'diffuseMap', diffuseMapDescriptor);
+visible(function () { /* Need to copy source code */ return false; })(SkyboxInfoProto, 'diffuseMap', diffuseMapDescriptor);
 const skyboxMaterialDescriptor = Object.getOwnPropertyDescriptor(SkyboxInfoProto, 'skyboxMaterial');
 tooltip('i18n:skybox.material')(SkyboxInfoProto, 'skyboxMaterial', skyboxMaterialDescriptor);
 type(Material)(SkyboxInfoProto, 'skyboxMaterial', skyboxMaterialDescriptor);
@@ -455,7 +455,7 @@ serializable(AmbientInfoProto, '_groundAlbedoLDR');
 const skyLightingColorDescriptor = Object.getOwnPropertyDescriptor(AmbientInfoProto, 'skyLightingColor');
 tooltip('i18n:ambient.skyLightingColor')(AmbientInfoProto, 'skyLightingColor', skyLightingColorDescriptor);
 editable(AmbientInfoProto, 'skyLightingColor', skyLightingColorDescriptor);
-visible(() => { /* Need to copy source code */ })(AmbientInfoProto, 'skyLightingColor', skyLightingColorDescriptor);
+visible(() => { /* Need to copy source code */ return false; })(AmbientInfoProto, 'skyLightingColor', skyLightingColorDescriptor);
 const skyIllumDescriptor = Object.getOwnPropertyDescriptor(AmbientInfoProto, 'skyIllum');
 tooltip('i18n:ambient.skyIllum')(AmbientInfoProto, 'skyIllum', skyIllumDescriptor);
 type(CCFloat)(AmbientInfoProto, 'skyIllum', skyIllumDescriptor);
@@ -463,5 +463,5 @@ editable(AmbientInfoProto, 'skyIllum', skyIllumDescriptor);
 const groundLightingColorDescriptor = Object.getOwnPropertyDescriptor(AmbientInfoProto, 'groundLightingColor');
 tooltip('i18n:ambient.groundLightingColor')(AmbientInfoProto, 'groundLightingColor', groundLightingColorDescriptor);
 editable(AmbientInfoProto, 'groundLightingColor', groundLightingColorDescriptor);
-visible(() => { /* Need to copy source code */ })(AmbientInfoProto, 'groundLightingColor', groundLightingColorDescriptor);
+visible(() => { /* Need to copy source code */ return false; })(AmbientInfoProto, 'groundLightingColor', groundLightingColorDescriptor);
 ccclass('cc.AmbientInfo')(AmbientInfo);
