+++ cocos/core/splash-screen.ts
@@ -39,8 +39,8 @@ import { Mat4, Vec2 } from './math';
 import { Settings, settings } from './settings';
 import { sys } from './platform/sys';
 
-// TODO: the type Image conflicts with the one on OH platform. 
-declare const Image: any;
+// TODO: the type Image conflicts with the one on OH platform.
+type Image = any;
 
 const v2_0 = new Vec2();
 type SplashEffectType = 'NONE' | 'FADE-INOUT';
