+++ cocos/native-binding/impl.ts
@@ -24,7 +24,9 @@
  */
 import { sys } from "../core/platform/sys";
 import { NATIVE } from 'internal:constants';
-const globalJsb = globalThis.jsb ?? {};
+const globalJsb: any = globalThis.jsb ?? {};
+declare const ScriptNativeBridge: any;
+
 if( NATIVE ){
     Object.defineProperty(globalJsb, 'reflection', {
         get () {
